feat(water): 新增应用内喝水登记、容器称重及历史记录 | Add in-app water check-in, container weighing and history
* 类型(Type): feat,service,refactor,ui

* 功能(Feature): 新增应用内喝水确认页；支持直接输入饮水容量，或选择饮水容器并根据总重量与空重自动换算毫升数；预置空重 241.5g 的常用水杯；新增包含本地图片的喝水历史记录

* 服务(Service): 调整喝水提醒通知及悬浮提醒跳转，携带提醒会话和测试状态进入应用内页面；新增容器、饮水记录、图片读取以及数据导入导出接口

* 重构(Refactor): 拆分应用导航、容器存储、喝水记录和迁移档案模块；采用稳定 ID、相对图片文件名及与包名无关的版本化数据格式，降低业务与应用包名、绝对路径之间的耦合

* 交互与界面(UI/UX): 新增已喝/未喝选择、容量与容器称重切换、实时饮水量换算、自拍选择、强制状态验证、容器管理及带图片的历史记录页面

* 文档(Docs): N/A

* Types: feat,service,refactor,ui

* Feature: Add an in-app water check-in page; support direct volume entry or automatic milliliter calculation from total weight and selected container tare weight; include a default 241.5g cup; add local water history with photos

* Service: Route water reminder notifications and overlays to the in-app flow with reminder session and test-state parameters; add APIs for containers, check-in records, local photos, and portable data import/export

* Refactor: Separate app navigation, container storage, water records, and migration archives; use stable IDs, relative photo filenames, and a versioned package-independent data format to reduce coupling

* UI/UX: Add drank/not-drank actions, volume and weighing modes, real-time calculation, selfie selection, forced state verification, container management, and photo-enabled history views

* Docs: N/A

+++ android/app/src/main/java/com/randomwaterreminder/app/ReminderAlertActivity.kt
@@ -405,7 +405,7 @@ class ReminderAlertActivity : Activity() {
             photo == null || !photo.isFile || photo.length() <= 0L -> Toast.makeText(this, "请先拍摄喝水自拍。", Toast.LENGTH_SHORT).show()
             else -> {
                 interactionHandled = true
-                if (!WaterCheckInStore.recordDrank(this, sessionId, amount, photo)) {
+                if (!WaterCheckInStore.recordDrank(this, sessionId, amount.toDouble(), photo)) {
                     interactionHandled = false
                     Toast.makeText(this, "本次记录未保存，可能已经处理过。", Toast.LENGTH_SHORT).show()
                     return
